Small refinement of the Ajax handlers
- included recipient_id to the MySQL delete statement
- some code clean up

+++ code/delete_postits.php
@@ -1,8 +1,8 @@
 <?php
 /*
- * Page module: PostIts
+ * Page module: Postits
  *
- * This module allows you to send virtual PostIts (sticky notes) to other users.
+ * This module allows you to send virtual Postits (sticky notes) to other users.
  * Requires some modification in the index.php file of the template.
  *
  * This file deletes a Postit from the database.
@@ -12,7 +12,7 @@
  * @platform    CMS WebsiteBaker 2.8.x
  * @package     postits
  * @author      cwsoft (http://cwsoft.de)
- * @version     1.2.0
+ * @version     1.2.1
  * @copyright   cwsoft
  * @license     http://www.gnu.org/licenses/gpl-3.0.html
 */
@@ -26,16 +26,26 @@ $admin = new admin('Pages', 'start', false, false);
 if (!$admin->is_authenticated()) return false;
 
 // check if a valid action is defined
-if (!(isset($_POST['action']) && $_POST['action'] == 'delete' && 
-	isset($_POST['postit_id']) && is_numeric($_POST['postit_id']) && isset($_POST['show']) && is_numeric($_POST['show']))) 
+if (! 
+	(
+	isset($_POST['action']) 
+	&& $_POST['action'] == 'delete' 
+	&& isset($_POST['postit_id']) 
+	&& is_numeric($_POST['postit_id']) 
+	&& isset($_POST['show']) 
+	&& is_numeric($_POST['show'])
+	)
+) {
 	return false;
-
+}
+	
 /*
- * Delete Postit specified by $_POST['postit_id']
+ * Delete Postit specified by $_POST['postit_id'], belonging to logged in user
  */
 $table = TABLE_PREFIX . 'mod_postits';
-$max_posts = ($_POST['show'] > 0) ? (int) $_POST['show'] : '5';
-$user_id = $admin->add_slashes((int) $admin->get_session('USER_ID'));
-$sql = "DELETE FROM `$table` WHERE `id` = '" . $_POST['postit_id'] . "'";
+$user_id = $admin->get_session('USER_ID');
+$sql = "DELETE FROM `$table` WHERE `id` = '" . (int) $_POST['postit_id'] . "' AND `recipient_id` = '" . (int) $user_id . "'";
 $database->query($sql);
+
+// return status
 echo ($database->is_error()) ? '{"status": "error"}' : '{"status": "ok"}';
